Revert "Move ServerConfig class to sub-package"
This reverts commit 03ead4676a83bd22dc6936cd466bec4d905f6994.
