+++ graphql/src/main/java/com/scalar/db/graphql/server/GraphQlServer.java
@@ -2,7 +2,6 @@ package com.scalar.db.graphql.server;
 
 import com.scalar.db.api.DistributedStorageAdmin;
 import com.scalar.db.config.DatabaseConfig;
-import com.scalar.db.graphql.server.config.ServerConfig;
 import com.scalar.db.service.StorageFactory;
 import java.io.File;
 import java.util.concurrent.Callable;
